src/etl/requirement.txt
requests==2.32.5
google-cloud-bigquery==3.38.0
google-auth==2.42.0
google-cloud-bigquery-storage==2.34.0
pytest==8.4.2
